feat: rework plugin deletion

+++ Dalamud/Interface/Internal/Windows/PluginInstaller/PluginInstallerWindow.cs
@@ -1675,6 +1675,12 @@ namespace Dalamud.Interface.Internal.Windows.PluginInstaller
                 trouble = true;
             }
 
+            // Scheduled for deletion
+            if (plugin.Manifest.ScheduledForDeletion)
+            {
+                label += Locs.PluginTitleMod_ScheduledForDeletion;
+            }
+
             ImGui.PushID($"installed{index}{plugin.Manifest.InternalName}");
             var hasChangelog = !plugin.Manifest.Changelog.IsNullOrEmpty();
 
@@ -2084,8 +2090,10 @@ namespace Dalamud.Interface.Internal.Windows.PluginInstaller
 
             var pluginManager = Service<PluginManager>.Get();
 
+            var devNotDeletable = plugin.IsDev && plugin.State != PluginState.Unloaded && plugin.State != PluginState.DependencyResolutionFailed;
+
             ImGui.SameLine();
-            if (plugin.State != PluginState.Unloaded && plugin.State != PluginState.LoadError)
+            if (plugin.State == PluginState.Loaded || devNotDeletable)
             {
                 ImGui.PushFont(InterfaceManager.IconFont);
                 ImGuiComponents.DisabledButton(FontAwesomeIcon.TrashAlt.ToIconString());
@@ -2093,18 +2101,9 @@ namespace Dalamud.Interface.Internal.Windows.PluginInstaller
 
                 if (ImGui.IsItemHovered())
                 {
-                    ImGui.SetTooltip(Locs.PluginButtonToolTip_DeletePluginLoaded);
-                }
-            }
-            else if (plugin.HasEverStartedLoad && !plugin.IsDev)
-            {
-                ImGui.PushFont(InterfaceManager.IconFont);
-                ImGuiComponents.DisabledButton(FontAwesomeIcon.TrashAlt.ToIconString());
-                ImGui.PopFont();
-
-                if (ImGui.IsItemHovered())
-                {
-                    ImGui.SetTooltip(Locs.PluginButtonToolTip_DeletePluginRestricted);
+                    ImGui.SetTooltip(plugin.State == PluginState.Loaded
+                                         ? Locs.PluginButtonToolTip_DeletePluginLoaded
+                                         : Locs.PluginButtonToolTip_DeletePluginRestricted);
                 }
             }
             else
@@ -2113,8 +2112,19 @@ namespace Dalamud.Interface.Internal.Windows.PluginInstaller
                 {
                     try
                     {
-                        plugin.DllFile.Delete();
-                        pluginManager.RemovePlugin(plugin);
+                        if (plugin.IsDev)
+                        {
+                            plugin.DllFile.Delete();
+                        }
+                        else
+                        {
+                            plugin.ScheduleDeletion(!plugin.Manifest.ScheduledForDeletion);
+                        }
+
+                        if (plugin.State is PluginState.Unloaded or PluginState.DependencyResolutionFailed)
+                        {
+                            pluginManager.RemovePlugin(plugin);
+                        }
                     }
                     catch (Exception ex)
                     {
@@ -2128,7 +2138,21 @@ namespace Dalamud.Interface.Internal.Windows.PluginInstaller
 
                 if (ImGui.IsItemHovered())
                 {
-                    ImGui.SetTooltip(Locs.PluginButtonToolTip_DeletePlugin);
+                    string tooltipMessage;
+                    if (plugin.Manifest.ScheduledForDeletion)
+                    {
+                        tooltipMessage = Locs.PluginButtonToolTip_DeletePluginScheduledCancel;
+                    }
+                    else if (plugin.State is PluginState.Unloaded or PluginState.DependencyResolutionFailed)
+                    {
+                        tooltipMessage = Locs.PluginButtonToolTip_DeletePlugin;
+                    }
+                    else
+                    {
+                        tooltipMessage = Locs.PluginButtonToolTip_DeletePluginScheduled;
+                    }
+
+                    ImGui.SetTooltip(tooltipMessage);
                 }
             }
         }
@@ -2465,6 +2489,8 @@ namespace Dalamud.Interface.Internal.Windows.PluginInstaller
 
             public static string PluginTitleMod_OrphanedError => Loc.Localize("InstallerOrphanedError", " (unknown repository)");
 
+            public static string PluginTitleMod_ScheduledForDeletion => Loc.Localize("InstallerScheduledForDeletion", " (scheduled for deletion)");
+
             public static string PluginTitleMod_New => Loc.Localize("InstallerNewPlugin ", " New!");
 
             #endregion
@@ -2539,6 +2565,10 @@ namespace Dalamud.Interface.Internal.Windows.PluginInstaller
             public static string PluginButtonToolTip_DeletePlugin => Loc.Localize("InstallerDeletePlugin ", "Delete plugin");
 
             public static string PluginButtonToolTip_DeletePluginRestricted => Loc.Localize("InstallerDeletePluginRestricted", "Cannot delete right now - please restart the game.");
+
+            public static string PluginButtonToolTip_DeletePluginScheduled => Loc.Localize("InstallerDeletePluginScheduled", "Delete plugin on next restart");
+
+            public static string PluginButtonToolTip_DeletePluginScheduledCancel => Loc.Localize("InstallerDeletePluginScheduledCancel", "Cancel scheduled deletion");
 
             public static string PluginButtonToolTip_DeletePluginLoaded => Loc.Localize("InstallerDeletePluginLoaded", "Disable this plugin before deleting it.");
